requirements.txt
asgiref==3.4.1
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.6
django-cors-headers==3.8.0
django_filter==22.1
django-phonenumber-field==6.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.41
Pillow==9.0.0
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
drf_spectacular~=0.24.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.8
